Ability to parametrize the Dexie query.
I'm not sure about the name, `createDexieArrayQueryWithSource`,
and actually not sure if it should be a separate function from
`createDexieArrayQuery`. It might be better to unify them to a single
interface.

I've been using code very close to this in a side-project for a while
(just in a private fork of solid-dexie), and at least in that context
it's worked well so far!